+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 * Added more natural attacks to monster parsing.
 * Better support for spaces in file paths.
 * Items copied from SRD now include DDB metadata for use when generating an adventure muncher config file.
+* Cobalt Cookie can now be checked on the settings page.
+* Campaigns are now detected based on Cobalt Cookie and drop down list presented.
 
 # 2.1.5